backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "minimum-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi[all]>=0.115.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic[email]>=2.10.0",
    "pydantic-settings>=2.6.0",
    "argon2-cffi>=23.1.0",
    "PyJWT>=2.10.0",
    "cryptography>=43.0.0",
    "redis[hiredis]>=5.2.0",
    "aiobotocore>=2.15.0",
    "structlog>=24.4.0",
    "slowapi>=0.1.9",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "httpx>=0.28.0",
    "python-multipart>=0.0.18",
    "uvicorn[standard]>=0.32.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "fakeredis[aioredis]>=2.26.0",
    "factory-boy>=3.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["app/tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "S"]
ignore = ["S101"]
